Update TrackStatistics.java
A code analysis program SonarQube detected 7 code smells in the constructor "TrackStatistics". The problem proposes "Change the name of this local variable to match the regular expression' [a-z][a-zA-Z0-9]*$'." These bugs are raised in order to address code smell and enhance code quality by following naming rules. The recommended field name should adhere to the naming rules while also improving the readability and maintainability of the code. Please adjust the field name and make the appropriate modifications to the code to address this issue.