Add settings toggles for chart series
Adds toggles to the User Interface section of the app Settings that
allow the user to display or hide the three series available in the
chart: Elevation, Speed/Pace, and Heart Rate. As a user of a heart rate
monitor primarily during strength workouts and combat sports, the
distance and elevation graphs are not helpful to me - in addition, their
scale means that I am not able to easily see my exact heart rate. This
should allow users in a similar situation to better determine exactly
what is most helpful for them to be seen on the graph.

Also updates heart rate series possible intervals - the smallest interval
was 25, which was too low for most zone 1 cardio charts - if your heart
rate goes from 70 to 125, with a 25 interval, most of the chart will be
whitespace.